requirements.txt
telethon
dotenv
pillow
textwrap